# server/openapi.json
{
  "components": {
    "schemas": {
...
      "CatalogModel": {
        "description": "A catalog entry: the OpenAI model fields plus the operator's favorite flag\nand the custom-model marker Node adds in `MergeCustomModels`.",
        "properties": {
          "custom": {
            "type": [
              "boolean",
              "null"
            ]
          },
          "favorite": {
            "type": "boolean"
          },
          "id": {
            "type": "string"
          },
          "object": {
            "type": "string"
          },
          "owned_by": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "object",
          "owned_by",
          "favorite"
        ],
        "title": "CatalogModel",
        "type": "object"
      },
...
      "ModelListResponse": {
        "properties": {
          "data": {
            "items": {
              "$ref": "#/components/schemas/CatalogModel"
            },
            "type": "array"
          },
          "object": {
            "type": "string"
          }
        },
        "required": [
          "object",
          "data"
        ],
        "title": "ModelListResponse",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "adminSession": {
        "description": "The admin session cookie. It also satisfies API-key auth on the routes that accept it.",
        "in": "cookie",
        "name": "srouter_admin_session",
        "type": "apiKey"
      },
      "bearerAuth": {
        "description": "The same API key in the `Authorization` header.",
        "scheme": "bearer",
        "type": "http"
      },
      "srouterApiKey": {
        "description": "An SRouter API key. `Authorization: Bearer <key>` carries the same credential.",
        "in": "header",
        "name": "x-api-key",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "description": "Multi-provider OpenAI and Anthropic compatible LLM gateway, as served by the Rust build (`server/`). One listener on `PORT` carries the API, the dashboard SPA, and the provider callbacks; there is no secondary OAuth listener. Every response uses snake_case field names (owner ruling 2026-10-07), and errors use the `{error:{message,type,code?,param?}}` envelope. Cloudflare Tunnel routes and `/v1/settings/fallbacks/*` are excluded from this build by owner ruling; they do not appear below and answer `404`.",
    "title": "SRouter API",
    "version": "0.2.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/models": {
      "get": {
        "parameters": [
          {
            "description": "Force a refresh of the cached catalog before answering.",
            "in": "query",
            "name": "refresh",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "Force a refresh of the cached catalog before answering.",
            "in": "query",
            "name": "force",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModelListResponse"
                }
              }
            },
            "description": "The allowlist-filtered catalog; hidden models and the models of disabled providers are dropped."
          }
        },
        "security": [
          {
            "srouterApiKey": []
          },
          {
            "bearerAuth": []
          },
          {
            "adminSession": []
          }
        ],
        "summary": "List the model catalog"
      },
      "post": {
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "properties": {
                  "favorite": {
                    "type": "boolean"
                  },
                  "hidden": {
                    "type": "boolean"
                  },
                  "model_id": {
                    "type": "string"
                  }
                },
                "required": [
                  "model_id"
                ],
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CatalogModel"
                }
              }
            },
            "description": "The model was already registered; the entry is returned."
          },
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CatalogModel"
                }
              }
            },
            "description": "The model was registered."
          }
        },
        "security": [
          {
            "adminSession": []
          }
        ],
        "summary": "Register a custom model"
      }
    },
    "/v1/models/{model}": {
      "delete": {
        "parameters": [
          {
            "description": "Percent-encoded model id.",
            "in": "path",
            "name": "model",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "deleted": {
                      "type": "boolean"
                    }
                  },
                  "required": [
                    "deleted"
                  ],
                  "type": "object"
                }
              }
            },
            "description": "The custom model was removed; a model that was not custom answers `404`."
          }
        },
        "security": [
          {
            "adminSession": []
          }
        ],
        "summary": "Remove a custom model"
      },
      "get": {
        "parameters": [
          {
            "description": "Percent-encoded model id; the catch-all segment accepts an id that contains `/`.",
            "in": "path",
            "name": "model",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CatalogModel"
                }
              }
            },
            "description": "One catalog entry."
          }
        },
        "security": [
          {
            "srouterApiKey": []
          },
          {
            "bearerAuth": []
          },
          {
            "adminSession": []
          }
        ],
        "summary": "Read one model"
      },
      "patch": {
        "parameters": [
          {
            "description": "Percent-encoded model id.",
            "in": "path",
            "name": "model",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "properties": {
                  "favorite": {
                    "type": "boolean"
                  },
                  "hidden": {
                    "type": "boolean"
                  }
                },
                "type": "object"
              }
            }
          },
          "required": false
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CatalogModel"
                }
              }
            },
            "description": "The entry with its new flags."
          }
        },
        "security": [
          {
            "adminSession": []
          }
        ],
        "summary": "Set the favorite or hidden flag"
      },
      "put": {
        "parameters": [
          {
            "description": "Percent-encoded model id.",
            "in": "path",
            "name": "model",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "properties": {
                  "favorite": {
                    "type": "boolean"
                  },
                  "hidden": {
                    "type": "boolean"
                  }
                },
                "type": "object"
              }
            }
          },
          "required": false
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CatalogModel"
                }
              }
            },
            "description": "The stored entry."
          }
        },
        "security": [
          {
            "adminSession": []
          }
        ],
        "summary": "Upsert a custom model"
      }
    },
...
    "/v1/v1/models": {
      "get": {
        "parameters": [
          {
            "description": "Force a refresh of the cached catalog before answering.",
            "in": "query",
            "name": "refresh",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "Force a refresh of the cached catalog before answering.",
            "in": "query",
            "name": "force",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModelListResponse"
                }
              }
            },
            "description": "The allowlist-filtered catalog; hidden models and the models of disabled providers are dropped."
          }
        },
        "security": [
          {
            "srouterApiKey": []
          },
          {
            "bearerAuth": []
          },
          {
            "adminSession": []
          }
        ],
        "summary": "List the model catalog"
      }
    },
    "/v1/v1/models/{model}": {
      "get": {
        "parameters": [
          {
            "description": "Percent-encoded model id; the catch-all segment accepts an id that contains `/`.",
            "in": "path",
            "name": "model",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CatalogModel"
                }
              }
            },
            "description": "One catalog entry."
          }
        },
        "security": [
          {
            "srouterApiKey": []
          },
          {
            "bearerAuth": []
          },
          {
            "adminSession": []
          }
        ],
        "summary": "Read one model"
      }
    }
  }
}